// src/models/legacy/actions/Actions.ts
export interface BaseCondition {
 /**
   * The main data source. Either the request or the response object.
   * This is required when the `source` is not equal to `value. In this case it is ignored
   */
  type?: ActionType;
  /**
   * Source of the data.
   */
  source: RequestDataSourceEnum | ResponseDataSourceEnum | 'value';
  /**
   * The path to the data. When `iteratorEnabled` is set then this
   * is a path counting from an array item. When not set an entire value of `source` is used.
   */
  path?: string;
  /**
   * This is only used when `source` is set to `value`. The data is not extracted from any of the request fields but this value is used.
   */
  value?: string;
}
/**
 * A configuration that extracts complex data from arrays.
 * @deprecated
 */
export interface DataSourceConfiguration extends BaseCondition {
  /**
   * When set the iterator configuration is enabled
   */
  iteratorEnabled?: boolean;
  /**
   * Array search configuration.
   */
  iterator?: IteratorConfiguration;
}

/**
 * Describes action's condition configuration.
 * @deprecated
 */
export interface Condition extends DataSourceConfiguration {
  /**
   * The value to compare to the result of extracted from the data source value.
   * Usually it is a string. For `statuscode` acceptable value is a number.
   */
  predictedValue?: string | number;
  /**
   * The comparison operator.
   */
  operator?: OperatorEnum;
  /**
   * Whether the condition always pass.
   * The condition is not really checked, values can be empty. The condition check always returns `true`.
   */
  alwaysPass?: boolean;
  /**
   * Options related to the UI state in the application.
   */
  view?: ConditionViewOptions;
}

/**
 * @deprecated
 */
export interface ConditionViewOptions {
  /**
   * Whether the condition editor is rendered in the "full" view
   * instead of the summary.
   */
  opened?: boolean;
}
// ...
export interface IteratorConfiguration {
  /**
   * The path to the property to use in the comparison.
   */
  path: string;
  /**
   * The value of the condition.
   */
  condition: string;
  /**
   * The comparison operator.
   */
  operator: OperatorEnum;
}

/**
 * @deprecated
 */
export type OperatorEnum = "equal" | "not-equal" | "greater-than" | "greater-than-equal" | "less-than" | "less-than-equal" | "contains" | "regex";